+++ src/extension.ts
@@ -216,9 +216,10 @@ export async function activate(context: vscode.ExtensionContext): Promise<ZoweEx
         vscode.commands.registerCommand("zowe.pasteDataSet", (node) => pasteDataSet(node, datasetProvider));
         vscode.commands.registerCommand("zowe.renameDataSetMember", (node) => datasetProvider.rename(node));
         vscode.commands.registerCommand("zowe.hMigrateDataSet", (node) => hMigrateDataSet(node));
-        vscode.commands.registerCommand("zowe.deleteProfile", async (node) => deleteProfile(datasetProvider, ussFileProvider, jobsProvider, node));
+        vscode.commands.registerCommand("zowe.deleteProfile", async (node) =>
+            Profiles.getInstance().deleteProfile(datasetProvider, ussFileProvider, jobsProvider, node));
         vscode.commands.registerCommand("zowe.cmd.deleteProfile", async () =>
-        deleteProfile(datasetProvider, ussFileProvider, jobsProvider));
+            Profiles.getInstance().deleteProfile(datasetProvider, ussFileProvider, jobsProvider));
         vscode.workspace.onDidChangeConfiguration(async (e) => {
             datasetProvider.onDidChangeConfiguration(e);
         });
@@ -272,7 +273,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<ZoweEx
         vscode.commands.registerCommand("zowe.uss.saveSearch", async (node: IZoweUSSTreeNode) => ussFileProvider.saveSearch(node));
         vscode.commands.registerCommand("zowe.uss.removeSavedSearch", async (node: IZoweUSSTreeNode) => ussFileProvider.removeFavorite(node));
         vscode.commands.registerCommand("zowe.uss.deleteProfile", async (node) =>
-            deleteProfile(datasetProvider, ussFileProvider, jobsProvider, node));
+            Profiles.getInstance().deleteProfile(datasetProvider, ussFileProvider, jobsProvider, node));
         vscode.workspace.onDidChangeConfiguration(async (e) => {
             ussFileProvider.onDidChangeConfiguration(e);
         });
@@ -337,7 +338,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<ZoweEx
         vscode.commands.registerCommand("zowe.jobs.saveSearch", async (node) => jobsProvider.saveSearch(node));
         vscode.commands.registerCommand("zowe.jobs.removeSearchFavorite", async (node) => jobsProvider.removeFavorite(node));
         vscode.commands.registerCommand("zowe.jobs.deleteProfile", async (node) =>
-            deleteProfile(datasetProvider, ussFileProvider, jobsProvider, node));
+            Profiles.getInstance().deleteProfile(datasetProvider, ussFileProvider, jobsProvider, node));
         const theTreeView = jobsProvider.getTreeView();
         context.subscriptions.push(theTreeView);
         if (!ISTHEIA) {
@@ -1481,139 +1482,6 @@ function appendSuffix(label: string): string {
 export function getDocumentFilePath(label: string, node: IZoweTreeNode) {
     return path.join(DS_DIR, "/" + getProfile(node) + "/" + appendSuffix(label) );
 }
-
-export async function deleteProfile(
-    datasetTree: IZoweTree<IZoweDatasetTreeNode>, ussTree: IZoweTree<IZoweUSSTreeNode>,
-    jobsProvider: IZoweTree<IZoweJobTreeNode>, node?: IZoweNodeType) {
-
-        let deleteLabel: string;
-        let deletedProfile: IProfileLoaded;
-        if (!node){
-            deletedProfile = await Profiles.getInstance().deleteProfileCommand();
-            deleteLabel = deletedProfile.name;
-        } else {
-            deletedProfile = node.getProfile();
-            deleteLabel = node.getProfileName();
-        }
-        await Profiles.getInstance().deleteProfile(deletedProfile);
-
-        // Delete from Data Set Tree
-        datasetTree.mSessionNodes.forEach((sessNode) => {
-            if (sessNode.getProfileName() === deleteLabel) {
-                datasetTree.deleteSession(sessNode);
-            }
-        });
-
-        // Delete from Data Set Favorites
-        datasetTree.mFavorites.forEach((ses) => {
-            const findNode = ses.label.substring(1, ses.label.indexOf("]")).trim();
-            if (findNode === deleteLabel) {
-                datasetTree.removeFavorite(ses);
-            }
-        });
-        await datasetTree.refresh();
-
-        // Delete from USS Tree
-        ussTree.mSessionNodes.forEach((sessNode) => {
-            if (sessNode.getProfileName() === deleteLabel) {
-                ussTree.deleteSession(sessNode);
-            }
-        });
-
-        // Delete from USS Favorites
-        ussTree.mFavorites.forEach((ses) => {
-            const findNode = ses.label.substring(1, ses.label.indexOf("]")).trim();
-            if (findNode === deleteLabel) {
-                ussTree.removeFavorite(ses);
-            }
-        });
-        await ussTree.refresh();
-
-        // Delete from Jobs Tree
-        jobsProvider.mSessionNodes.forEach((jobNode) => {
-            if (jobNode.getProfileName() === deleteLabel) {
-                jobsProvider.deleteSession(jobNode);
-            }
-        });
-
-        // Delete from Jobs Favorites
-        jobsProvider.mFavorites.forEach((ses) => {
-            const findNode = ses.label.substring(1, ses.label.indexOf("]")).trim();
-            if (findNode === deleteLabel) {
-                jobsProvider.removeFavorite(ses);
-            }
-        });
-        await jobsProvider.refresh();
-}
-
-// export async function deleteProfileCommand(
-//     datasetTree: IZoweTree<IZoweDatasetTreeNode>, ussTree: IZoweTree<IZoweUSSTreeNode>, jobsProvider: IZoweTree<IZoweJobTreeNode>) {
-//         const sessionName = await Profiles.getInstance().deleteProfileCommand();
-
-//         const datasetNode = await datasetTree.mSessionNodes.find((tempNode) => tempNode.label.trim() === sessionName);
-//         if (datasetNode) {
-//             datasetTree.deleteSession(datasetNode);
-//         }
-//         datasetTree.mSessionNodes.forEach((sessNode) => {
-//             if (sessNode.contextValue === DS_SESSION_CONTEXT) {
-//                 utils.labelHack(sessNode);
-//                 sessNode.children = [];
-//                 sessNode.dirty = true;
-//                 utils.refreshTree(sessNode);
-//             }
-//         });
-
-//         // Delete from Data Set Favorites
-//         datasetTree.mFavorites.forEach((ses) => {
-//             const findNode = ses.label.substring(1, ses.label.indexOf("]")).trim();
-//             if (findNode === datasetNode.label) {
-//                 datasetTree.removeFavorite(ses);
-//             }
-//         });
-//         await datasetTree.refresh();
-
-//         const ussNode = await ussTree.mSessionNodes.find((tempNode) => tempNode.label.trim() === sessionName);
-//         if (ussNode) {
-//             ussTree.deleteSession(ussNode);
-//         }
-//         ussTree.mSessionNodes.forEach((sessNode) => {
-//             if (sessNode.contextValue === USS_SESSION_CONTEXT) {
-//                 utils.labelHack(sessNode);
-//                 sessNode.children = [];
-//                 sessNode.dirty = true;
-//                 utils.refreshTree(sessNode);
-//             }
-//         });
-//         // Delete from USS Favorites
-//         ussTree.mFavorites.forEach((ses) => {
-//             const findNode = ses.label.substring(1, ses.label.indexOf("]")).trim();
-//             if (findNode === ussNode.getProfileName()) {
-//                 ussTree.removeFavorite(ses);
-//             }
-//         });
-//         await ussTree.refresh();
-
-//         const jobsNode = await jobsProvider.mSessionNodes.find((tempNode) => tempNode.label.trim() === sessionName);
-//         if (jobsNode) {
-//             jobsProvider.deleteSession(jobsNode);
-//         }
-//         jobsProvider.mSessionNodes.forEach((jobNode) => {
-//             if (jobNode.contextValue === JOBS_SESSION_CONTEXT) {
-//                 utils.labelHack(jobNode);
-//                 jobNode.children = [];
-//                 jobNode.dirty = true;
-//                 utils.refreshTree(jobNode);
-//             }
-//         });
-//         // Delete from Jobs Favorites
-//         jobsProvider.mFavorites.forEach((ses) => {
-//             const findNode = ses.label.substring(1, ses.label.indexOf("]")).trim();
-//             if (findNode === jobsNode.getProfileName()) {
-//                 jobsProvider.removeFavorite(ses);
-//             }
-//         });
-//         await jobsProvider.refresh();
-// }
 
 export async function openRecentMemberPrompt(datasetTree: IZoweTree<IZoweDatasetTreeNode>, ussTree: IZoweTree<IZoweUSSTreeNode>) {
     if (log) {
